+++ resources/views/development.blade.php
@@ -12,6 +12,10 @@
 
                 <dl class="divide-y divide-gray-100 text-sm p-2">
                     <div class="p-2">
+                        <dt class="font-bold">App version</dt>
+                        <dd class="text-gray-700">{{config('app.version')}} - {{config('app.deployDateTime')}}</dd>
+                    </div>
+                    <div class="p-2">
                         <dt class="font-bold">Your IP</dt>
                         <dd class="text-gray-700">{{implode(', ', request()->ips())}}</dd>
                     </div>
@@ -70,22 +74,14 @@
             </script>
             <details open>
                 <summary>Web Push Notification Info</summary>
+                <div class="mx-2 mt-3">
+                    <span>Legend: </span>
+                    <i class='text-green-700 fa-solid fa-circle-check'></i> Passed
+                    <i class='text-orange-700 fa-solid fa-triangle-exclamation'></i> Not Checked
+                    <i class='text-red-700 fa-solid fa-circle-xmark'></i> Failed
+                </div>
                 <ol class="text-sm p-2 text-gray-700" x-data="initWebPush()"
                     x-on:webpush-setup-finished.window="updateFlags">
-                    <li>
-                        <i :class="{
-                        'text-green-700 fa-solid fa-circle-check': forceIsReady,
-                        'text-red-700 fa-solid fa-circle-xmark': forceIsReady === false}"></i> is browser ready
-                        <template x-if="forceIsReady"><span class="text-xs"
-                                                            x-text="'('+(new Date(Number(webPush.getLastCheckTimeStamp())))+')'"></span>
-                        </template>
-                    </li>
-                    <li>
-                        <i :class="{
-                        'text-green-700 fa-solid fa-circle-check': isBrowserReady,
-                        'text-orange-700 fa-solid fa-triangle-exclamation': isBrowserReady === null,
-                        'text-red-700 fa-solid fa-circle-xmark': isBrowserReady === false}"></i> is browser readydat
-                    </li>
                     <li>
                         <i :class="{
                         'text-green-700 fa-solid fa-circle-check': Notification.permission === 'granted',
@@ -93,6 +89,32 @@
                         'text-red-700 fa-solid fa-circle-xmark': Notification.permission === 'denied'}"></i> has
                         notification permission
                         (<span class="text-gray-500" x-text="Notification.permission"></span>)
+                    </li>
+                    <li x-data="{serviceWorkerSupported: ('serviceWorker' in navigator)}">
+                        <i :class="{
+                        'text-green-700 fa-solid fa-circle-check': serviceWorkerSupported,
+                        'text-red-700 fa-solid fa-circle-xmark': serviceWorkerSupported === false}"></i> are service
+                        worker supported by your browser
+                    </li>
+                    <li x-data="{pushManagerSupported: ('PushManager' in window)}">
+                        <i :class="{
+                        'text-green-700 fa-solid fa-circle-check': pushManagerSupported,
+                        'text-red-700 fa-solid fa-circle-xmark': pushManagerSupported === false}"></i> is the push
+                        manager supported by your browser
+                    </li>
+                    <li>
+                        <i :class="{
+                        'text-green-700 fa-solid fa-circle-check': isBrowserReady,
+                        'text-orange-700 fa-solid fa-triangle-exclamation': isBrowserReady === null,
+                        'text-red-700 fa-solid fa-circle-xmark': isBrowserReady === false}"></i> is browser ready
+                    </li>
+                    <li>
+                        <i :class="{
+                        'text-green-700 fa-solid fa-circle-check': forceIsReady,
+                        'text-red-700 fa-solid fa-circle-xmark': forceIsReady === false}"></i> checked is cached
+                        <template x-if="forceIsReady"><span class="text-xs"
+                                                            x-text="'('+(new Date(Number(webPush.getLastCheckTimeStamp())))+')'"></span>
+                        </template>
                     </li>
                     <li>
                         <i :class="{
@@ -115,10 +137,13 @@
                         subscription stored
                     </li>
                     <li>
-                        <span>is vapid public key stored ?</span> - <span class="text-gray-500"
-                                                                          x-text="webPush.isVapidPublicKeyStored() ? 'yes': 'no'"></span>
-                        <template x-if="webPush.isVapidPublicKeyStored"><span class="text-xs"
-                                                                              x-text="webPush.getStoredVapidPublicKey()"></span>
+                        <i x-data="{isVapidKeyStored: webPush.isVapidPublicKeyStored()}" :class="{
+                        'text-green-700 fa-solid fa-circle-check': isVapidKeyStored,
+                        'text-orange-700 fa-solid fa-triangle-exclamation': isVapidKeyStored === null,
+                        'text-red-700 fa-solid fa-circle-xmark': isVapidKeyStored === false}"></i>
+                        <span>is vapid public key stored</span>
+                        <template x-if="webPush.isVapidPublicKeyStored">
+                            <div class="text-xs ml-5 break-all" x-text="webPush.getStoredVapidPublicKey()"></div>
                         </template>
                     </li>
                 </ol>
@@ -126,21 +151,43 @@
         </div>
         <div class="shadow-xl shadow-black/5 sm:rounded-md bg-white p-3">
             <details open>
-                <summary>Global DB configuration settings</summary>
+                <summary>PHP (env) config</summary>
                 <ol class="text-sm p-2 text-gray-700">
-                    @foreach([] as $config)
-                        <li><span>{{$config->key}}</span><span>{{$config->value}}</span></li>
+                    @foreach(['app.name', 'app.env', 'app.debug', 'app.url', 'app.displayed_timezone',
+                                'app.locale', 'app.currency', 'logging.default', 'session.driver',
+                                'session.lifetime','session.expire_on_close', 'session.encrypt'] as $configKey)
+                        <li>
+                            <span class="font-bold">{{$configKey}}:</span>
+                            <pre class="inline">"{{config($configKey)}}"</pre>
+                        </li>
                     @endforeach
                 </ol>
             </details>
         </div>
-        @if()
+        <div class="shadow-xl shadow-black/5 sm:rounded-md bg-white p-3">
+            <details open>
+                <summary>Global DB configuration settings</summary>
+                <ol class="text-sm p-2 text-gray-700">
+                    @foreach(\App\Models\Configuration::query()->whereNull('user_id')->get() as $config)
+                        <li>
+                            <span class="font-bold">{{\App\Models\ConfigurationKey::from($config->key)->name}}:</span>
+                            <pre class="inline">"{{$config->value}}"</pre>
+                        </li>
+                    @endforeach
+                </ol>
+            </details>
+        </div>
+        @if($id = auth()->id())
             <div class="shadow-xl shadow-black/5 sm:rounded-md bg-white p-3">
                 <details open>
                     <summary>Your user specific DB configuration settings</summary>
                     <ol class="text-sm p-2 text-gray-700">
-                        @foreach()
-                            <li><span>{{$key}}</span><span>{{$value}}</span></li>
+                        @foreach(\App\Models\Configuration::query()->where('user_id', $id)->get() as $config)
+                            <li>
+                                <span
+                                    class="font-bold">{{\App\Models\ConfigurationKey::from($config->key)->name}}:</span>
+                                <pre class="inline">"{{$config->value}}"</pre>
+                            </li>
                         @endforeach
                     </ol>
                 </details>