+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Added
+- add development page with different details, and a config to enable (public) access to it
 
 ## v0.7.6 - 2024-04-08
 ### Added